+++ packages/mui-components/src/index.ts
@@ -3,5 +3,4 @@ export * from './form';
 export * from './theme';
 export * from './functions';
 export * from './hooks';
-export * from './test-utils';
 export * from './types';