+++ packages/mui-components/src/theme/components/MuiInputBase/index.ts
@@ -3,35 +3,23 @@ import type { Components, Theme } from '@mui/material';
 export const MuiInputBase: Components<Theme>['MuiInputBase'] = {
   styleOverrides: {
     root: ({ theme }) => ({
-      'border': `1px solid ${theme.palette.divider}`,
       'padding': '0rem 1rem',
-      'transition': '0.3s ease-out',
-
-      '&:hover': {
-        background: theme.designTokens.stateLayers.interactiveSurface,
-      },
-
-      '&:focus-within': {
-        background: theme.designTokens.stateLayers.interactiveSurface,
-      },
+      'transition': 'background-color 150ms ease, border-color 150ms ease, box-shadow 150ms ease',
 
       '& .MuiOutlinedInput-notchedOutline': {
         border: 'none',
       },
 
       '&.Mui-error': {
-        'borderColor': theme.palette.error.main,
-        'borderWidth': '2px',
         'outlineOffset': '3px',
 
         '& .MuiInputAdornment-root': {
-          height: 'calc(3rem - 4px)',
+          height: 'calc(3rem - 2px)',
         },
       },
 
       '&.Mui-disabled': {
         backgroundColor: theme.palette.action.disabledBackground,
-        border: `1px solid ${theme.palette.divider}`,
       },
 
       '&[type=number]': {